script_titan/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
psycopg2-binary
python-dotenv
matplotlib